Test some functions of queue
I modified some functions in queue.c, which are new, free, insert
, remove, size and delete_mid.

In the remove operation, I encountered the problem. There are two
ways to unlink the element from the list. The one with initialization
could not pass the complexity test, even though they only differ by
the constants.

I thought this is the environment issue, because I test this lab
in windows subsystem linux. Maybe I should change my workspace to
linux or virtual machine.

This is the my first try to commit. I did lots things but I'm too
lazy to write the detail of the implement in each functions. I
will practice this principle to the latter commit.

Change-Id: I57f22b392c18a4c737322eec12af588bd239214b