Add the swap operation
I use the indirect method to implement this operation.

We just focus on the pointer to the next node. If it
is the head or the previous one of the head, the process
is done; otherwise, we need to change the next two nodes.

This pointer can avoid us from messing up with the order.

Change-Id: I3282e02b527f312123f89d9dc436f26e9e97063b