Add ascend/descend and modify insert operation
In ascend/descend operation, using the technique called monotone stack
can make the complexity be linear. In implementation, we can keep track
of the element that holds the smallest/largest value, while traversing
backward. Notice that the length of value is bounded in the insert
operation, the comparison between the values would always success.

On the other sides, in the insert operation, there is a mistake. When
the memory allocation fails for the value, the original code doesn't
release the memory of element. I fixed that issue.

Change-Id: I3bdf8c65fd7b3b6a75e706a07c73bb0b4bb2b742