Add sort operation
I use merge sort as the sort function for queue. When it divide into
two queue, I borrow the memory from stack to make the temporary queue
heads. This method can avoid from memory allocation, which could fail
when the memory is not enough.

Also, I make the merge process into a function, which can be used by
merge operation for queue easily.

Change-Id: I920284d436c354db65ee97d37a7ae9c963a2c850