Modify reverseK and merge sort operation
In reverseK operation, We can use two instance to be temporary heads
and the usage of them are like stack.

In every iteration, we pop the first K elements from head to cache
, and the order of elements in the cache is reversed. Next we can
easily flush the cache to the dummy head. If the number of elements
in cache is less than K, we reverse the queue in the queue, then paste
it to the end of dummy head. Last, replace the dummy head with the
original one.

We have another way to implement this operation, which is the naive
thought. Using safe version traversal, cutting and reversing every
K elements. We can use defined function to do every step. However
the implementation of cut operation needs lots of branch instruction
, which might decrease the efficiency when k is small.

On the other hand, we can use cut operation to simplify the section
of merge sort that split the queue into two half parts.

Change-Id: Ia56b11ca3beb45e650a7b51d1447def9175aa922