Modify dudect library
Add percentile and cropping module like the one appears in the original
source code. Note that time context should be clear before test
starting; otherwise, t-value will increasing rapidly. Besides
, percentile should be zero and be prepared before the main test.
Therefore we need do extra one "doit" function for the percentile.

In the measure function, if we drop samples first, the result
statistics will contain zeros at both start and end. The correct
way is dropping the samples when updating statistics.

When we need to test the function, it is necessary that put the memory
fence instruction before we get the time stamp counter. All the load
and store function should complete before reading counter.

Last, I moved the position of memory allocation to the test function.
It is redundant and costly to use it in the "doit" function.

About the modification of list.h and checksum, I found that the new
one somehow cannot pass the static test, although I added suppress
hint before the macro. Maybe I should consult to the lecturer for help.

Change-Id: If14915992147f6099f52f703699b2c2e6f1da666